feat: update types with RawDessertProducts
